common.utils: automatic refactor: time
Move `nice_duration`, `get_timezone`, `parsetime`, `strtotime` and
`strtodate` to `common.time`.
