+++ common/highlights.py
@@ -1,3 +1,4 @@
+import common.time
 from common import utils
 
 SPREADSHEET = "1yrf6d7dPyTiWksFkhISqEc-JR71dxZMkUoYrX4BR40Y"
@@ -7,6 +8,6 @@ def format_row(title, description, url, timestamp, nick):
 		("SHOW", title),
 		("QUOTE or MOMENT", description),
 		("YOUTUBE VIDEO LINK", url),
-		("ROUGH TIME THEREIN", "before " + utils.nice_duration(timestamp, 0)),
+		("ROUGH TIME THEREIN", "before " + common.time.nice_duration(timestamp, 0)),
 		("NOTES", "from chat user '%s'" % nick),
 	]
